+++ common.h
@@ -13,27 +13,27 @@ struct task {
 
 	bool isOverloadTask;
 
-	float orig_nominal_exectime_ns;
-	float orig_period_ns;
+	double orig_nominal_exectime_ns;
+	double orig_period_ns;
 
-	float exectime_ns;
-	float nominal_exectime_ns;
-	float exectime_in_rm_ns;
-	float period_ns;
-	float criticality;
+	double exectime_ns;
+	double nominal_exectime_ns;
+	double exectime_in_rm_ns;
+	double period_ns;
+	double criticality;
 
 	struct task *overload_task;
 };
 
 
-float getResponseTimeCAPA(struct task *table, int tablesize, struct task *rtask);
+double getResponseTimeCAPA(struct task *table, int tablesize, struct task *rtask);
 struct task *parseArgs(int argc, char **argv, int *tablesize);
 
 bool OPTIsTaskSched(struct task *table, int numEntry, bool *checkPass);
 bool VOPTIsTaskSched(struct task *table, int numEntry, bool *checkPass);
 
 bool PTIsTaskSchedNaive(struct task *table, int numEntry, bool *checkPass);
-float getResponseTimeRM(struct task *table, int tablesize,
+double getResponseTimeRM(struct task *table, int tablesize,
 						struct task *rtask);
 int criticalitySort(const void *a, const void *b);
 void copyTask(struct task *dest, struct task *src);
@@ -43,7 +43,7 @@ bool CAPAIsTaskSched(struct task *table, int numEntry, bool *checkPass);
 bool PTIsTaskSched(struct task *table, int numEntry, bool *checkPass);
 bool ZSRMSIsTaskSched(struct task *table, int numEntry, bool *checkPass);
 
-float getResponseTimePT(struct task *table, int tablesize,
+double getResponseTimePT(struct task *table, int tablesize,
 				struct task *rtask);
 
 #endif /* __COMMON_H__ */
